#175 added desktop structure
